No static variable for search + split IntervalTree
Using a static variable for the search was a bit confusing and obfuscated what was really happening.
As it's easy to extract the data from records it now returns the records themselves instead of their data.
IntervalTree has been split into IntervalTree and the default export DataIntervalTree, adding an option for advanced users that need extra flexibility.
DataIntervalTree wraps IntervalTree to provide an easy to use interface, and to not break API-compatibility.

Use client created records

This is an attempt to simplify the API by exposing less internal data structures while keeping the client in control.

Add DataIntervalTree and make it default

DataIntervalTree makes it easier to get started as the call signature clearly shows what arguments it need.